backend/requirements.txt
fastapi
uvicorn
openai
faiss-cpu
python-dotenv
pydantic
httpx
python-multipart
aiofiles
PyPDF2
beautifulsoup4
requests
lxml
